Improve insertion and deletion (from end) time complexity from Average O(1), Worst case O(N^2) to Amortized O(1), Worst case O(N).
Previously, resizing involved adding a constant amount of new space to the new array and copying over the old
elements. This resulted in a worst case time complexity of O(N^2).

Proof:
Consider an array of initial size 1 with a resize scheme that every insertion requires resize of array.

- Insertion of first element takes 1 opertaion.
- Insertion of second element means resize the array, copy over the old element, then insert new element
  (2 operations - copy old element, insert new one)
- Insertion of third element - copy prev 2 elements, insert new one (3 operations).

In general, for n insertions, the operations can be listed as
    1 + 2 + 3 + 4 + 5 ........+ N
which can be summed as n(n+1)/2 = 1/2 n^2 + 1/2 n = O(n^2) worst case operations.

If the constant resize factor is increased, say we increase the size by 100 instead of 1 and
initial size of array is 100, the asymptotic analysis is no better. i.e.,

-Insertion of first 100 elements take constant time (100 operations)
-Insertion of next 100 elements means first copy over the prev 100 elements, then insert the new ones
 in constant time (200 operations)
-Insertion of next 100 elements means first copy over the prev 200 elements, then insert the new batch
 of 100 elements in constant time (300 operations)

In general for n insertions, the operations can be listed as:
    100 + 200 + 300 + 400 + ........+ N

which is an arithmetic series with sum given as:
Sn = n/2 (2a + (n - 1) d) = n/2 (2(100) + (n - 1) 100) = 50n + 50n^2 = O(n^2) worst case operations.

Even if we resize by K elements where K is a very large number, the asymptotic complexity remains
O(n^2) in the worst case. (And the case where K is a very large number is actually not desirable as
well as a large chunk of memory would remain empty at any given instant wasting a lot of memory.)

The solution is to scale the resize depending upon the size of the current array to reduce number of resizes.
If we resize the array by a factor of 2 each time (and not by a constant amount), and perform the same
analysis when initial array size is 1:

- Insert first element = 1 operation.
- Insert 2 elements = 2 operations (copy old + insertion)
- Insert 4 elements = 4 operations (2 for copying old elem, 2 for constant insertions after resize)
- Insert 8 elements = 8 operations (4 for copying old elem, 4 for constant insertions after resize)

In general for N insertions (assuming N is a power of two for simpler math), the trend is given as:
    1 + 2 + 4 + 8 +........+ N

    or we can write it as:

    N + N/2 + N/4 +........+ 1

which is geometric series whose sum is given as 2N - 1 = O(N) worst case insertions.
(For proof: https://math.stackexchange.com/questions/401937/how-is-nn-2n-4-1-equal-to-2n-1-using-the-formula-for-geometric-series)

Amortized O(1):
Due to the optimization of delete operation (detail given in line 133 of assignment1.cpp), we achieve Amortized O(1) instead of Average O(1).

Average O(1) vs Amortized O(1):
    Average O(1) means if we pick any operation 'at random', the cost of it would be in Big O(1)
    which is true as we are doing the cheaper operations far more frequently than the costly ones
    that require resize of array. However, average does not exclude the possibility in which one gets
    very 'unlucky' and we hit a case when we are inserting and deleting one after the other in such a manner
    that the insert requires growth of array and immiadate delete after it requires shirnk of array. In such
    case, the performance of the alogrithm will be much slower than what is being indicated by average case
    (which involves operations to be selected at random due to which this case is skipped more often in average
    calculations).

    Amortized O(1) is better than average O(1). It makes no assumptions if the inputs are random are not. It will
    consider the each and every case and does not require the inputs to be random. It considers the total performance
    of a sequence of operations instead of just one operation so we have a guarantee on the total cost of a sequence of operations.

Sources:
https://gist.github.com/jconnolly/5acf05f279a7e9e40371
https://stackoverflow.com/questions/7333376/difference-between-average-case-and-amortized-analysis
https://youtu.be/5AllG-i_yto
https://youtu.be/Ij7NQ-0mIVA